+++ libmproxy/protocol2/messages.py
@@ -1,6 +1,7 @@
 """
 This module contains all valid messages layers can send to the underlying layers.
 """
+from __future__ import (absolute_import, print_function, division, unicode_literals)
 
 
 class _Message(object):
